requirements.txt
numpy==1.26.4
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
yfinance>=0.2.40
finnhub-python>=2.4.19
transformers>=4.40.0
torch==2.4.0